# perf(mempool): Replace `sync.Mutex` with `sync.Once`